requirements.txt
flask==2.3.3
python-gitlab==3.15.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
gunicorn==21.2.0
redis==5.0.1 